spline #954 arangodb-foxx-services Producer REST to delegate to Foxx. Remove AppTxBuilder.
